requirements.txt
Flask
Werkzeug
Flask-Script
Flask-WTF
pyserial